+++ src/main/java/roomescape/admin/reservation/service/ReservationService.java
@@ -3,7 +3,7 @@ package roomescape.admin.reservation.service;
 import java.util.List;
 import org.springframework.stereotype.Service;
 import roomescape.admin.reservation.entity.Reservation;
-import roomescape.admin.reservation.repository.ReservationRepository;
+import roomescape.admin.reservation.service.port.ReservationRepository;
 
 @Service
 public class ReservationService {
